change int into short

+++ 6_2/6_2/源.cpp
@@ -7,19 +7,19 @@ using namespace std;
 class problem
 {
 private:
-	vector<vector<int>> graph;
-	vector<vector<int>> tgraph;
-	vector<int> stack;
+	vector<vector<short>> graph;
+	vector<vector<short>> tgraph;
+	vector<short> stack;
 	vector<bool> visited;
-	vector<vector<int>> scc;
-	vector<int> vertex_to_scc;
-	vector<int> scc_impact;
+	vector<vector<short>> scc;
+	vector<short> vertex_to_scc;
+	vector<short> scc_impact;
 private:
-	void first_round_dfs(int index);
-	void second_round_dfs(int index);
+	void first_round_dfs(short index);
+	void second_round_dfs(short index);
 	void get_t_graph();
-	void get_one_scc_impact(int scc_index);
-	int impact_dfs(int index);
+	void get_one_scc_impact(short scc_index);
+	short impact_dfs(short index);
 public:
 	problem();
 	void get_data();
@@ -35,13 +35,13 @@ problem::problem()
 void problem::get_data()
 {
 	string s_line = "";
-	int temp = 0;
+	short temp = 0;
 	while (getline(cin, s_line))
 	{
 		stringstream ss;
 		ss.clear();
 		ss << s_line;
-		vector<int> v_line;
+		vector<short> v_line;
 		while (!ss.eof())
 		{
 			ss >> temp;
@@ -55,11 +55,11 @@ void problem::get_data()
 void problem::get_t_graph()
 {
 	tgraph.resize(graph.size());
-	for (int i = 0; i < (int)graph.size(); i++)
+	for (short i = 0; i < (short)graph.size(); i++)
 	{
-		for (int j = 0; j < (int)graph[i].size(); j++)
+		for (short j = 0; j < (short)graph[i].size(); j++)
 		{
-			int dest = graph[i][j];
+			short dest = graph[i][j];
 			tgraph[dest].push_back(i);
 		}
 	}
@@ -69,7 +69,7 @@ void problem::get_scc()
 {
 	//first round dfs
 	fill(visited.begin(), visited.end(),false);
-	for (int i = 0; i < (int)graph.size(); i++)
+	for (short i = 0; i < (short)graph.size(); i++)
 	{
 		if (!visited[i])
 		{
@@ -87,17 +87,17 @@ void problem::get_scc()
 	fill(visited.begin(), visited.end(), false);
 	while (stack.size() > 0)
 	{
-		int vertex_index = stack[stack.size() - 1];
+		short vertex_index = stack[stack.size() - 1];
 		stack.pop_back();
 
 		if (!visited[vertex_index])
 		{
 			visited[vertex_index] = true;
 
-			vector<int> one_scc;
+			vector<short> one_scc;
 			one_scc.push_back(vertex_index);
 			scc.push_back(one_scc);
-			vertex_to_scc[vertex_index] = (int)(scc.size() - 1);
+			vertex_to_scc[vertex_index] = (short)(scc.size() - 1);
 
 			second_round_dfs(vertex_index);
 		}
@@ -107,11 +107,11 @@ void problem::get_scc()
 }
 
 
-void problem::first_round_dfs(int index)
+void problem::first_round_dfs(short index)
 {
-	for (int i = 0; i < (int)graph[index].size(); i++)
+	for (short i = 0; i < (short)graph[index].size(); i++)
 	{
-		int nei_index = graph[index][i];
+		short nei_index = graph[index][i];
 		if (!visited[nei_index])
 		{
 			visited[nei_index] = true;
@@ -121,16 +121,16 @@ void problem::first_round_dfs(int index)
 	}
 }
 
-void problem::second_round_dfs(int index)
+void problem::second_round_dfs(short index)
 {
-	for (int i = 0; i < (int)tgraph[index].size(); i++)
+	for (short i = 0; i < (short)tgraph[index].size(); i++)
 	{
-		int nei_index = tgraph[index][i];
+		short nei_index = tgraph[index][i];
 		if (!visited[nei_index])
 		{
 			visited[nei_index] = true;
 			scc[scc.size() - 1].push_back(nei_index);
-			vertex_to_scc[nei_index] = (int)(scc.size() - 1);
+			vertex_to_scc[nei_index] = (short)(scc.size() - 1);
 			second_round_dfs(nei_index);
 		}
 	}
@@ -139,12 +139,12 @@ void problem::second_round_dfs(int index)
 void problem::get_all_scc_impact()
 {
 	scc_impact.resize(scc.size(), -1);
-	for (int i = 0; i < (int)scc.size(); i++)
+	for (short i = 0; i < (short)scc.size(); i++)
 	{
 		get_one_scc_impact(i);
 	}
 }
-void problem::get_one_scc_impact(int scc_index)
+void problem::get_one_scc_impact(short scc_index)
 {
 	if (scc_impact[scc_index] >= 0)
 	{
@@ -157,12 +157,12 @@ void problem::get_one_scc_impact(int scc_index)
 		scc_impact[scc_index]=impact_dfs(scc[scc_index][0])-1;
 	}
 }
-int problem::impact_dfs(int index)
+short problem::impact_dfs(short index)
 {
-	int result = 1;//myself
-	for (int i = 0; i < (int)graph[index].size(); i++)
+	short result = 1;//myself
+	for (short i = 0; i < (short)graph[index].size(); i++)
 	{
-		int nei_index = graph[index][i];
+		short nei_index = graph[index][i];
 		if (!visited[nei_index])
 		{
 			visited[nei_index] = true;
@@ -182,8 +182,8 @@ int problem::impact_dfs(int index)
 void problem::output()
 {
 	cout << "I have read the rules about plagiarism punishment" << endl;
-	int max_impact = 0, max_index = 0;
-	for (int i = 0; i < (int)scc_impact.size(); i++)
+	short max_impact = 0, max_index = 0;
+	for (short i = 0; i < (short)scc_impact.size(); i++)
 	{
 		if (scc_impact[i] > max_impact)
 		{
@@ -193,7 +193,7 @@ void problem::output()
 	}
 	cout << max_impact << endl;
 	sort(scc[max_index].begin(), scc[max_index].end());
-	for (int i = 0; i < (int)scc[max_index].size(); i++)
+	for (short i = 0; i < (short)scc[max_index].size(); i++)
 	{
 		cout << scc[max_index][i] << " ";
 	}
